+++ back/src/Services/SocketManager.ts
@@ -701,8 +701,8 @@ export class SocketManager {
             return;
         }
 
-        const recipient = room.getUserByUuid(recipientUuid);
-        if (recipient === undefined) {
+        const recipients = room.getUsersByUuid(recipientUuid);
+        if (recipients.length === 0) {
             console.error(
                 "In sendAdminMessage, could not find user with id '" +
                     recipientUuid +
@@ -711,14 +711,16 @@ export class SocketManager {
             return;
         }
 
-        const sendUserMessage = new SendUserMessage();
-        sendUserMessage.setMessage(message);
-        sendUserMessage.setType("ban"); //todo: is the type correct?
+        for (const recipient of recipients) {
+            const sendUserMessage = new SendUserMessage();
+            sendUserMessage.setMessage(message);
+            sendUserMessage.setType("ban"); //todo: is the type correct?
 
-        const serverToClientMessage = new ServerToClientMessage();
-        serverToClientMessage.setSendusermessage(sendUserMessage);
+            const serverToClientMessage = new ServerToClientMessage();
+            serverToClientMessage.setSendusermessage(sendUserMessage);
 
-        recipient.socket.write(serverToClientMessage);
+            recipient.socket.write(serverToClientMessage);
+        }
     }
 
     public async banUser(roomId: string, recipientUuid: string, message: string): Promise<void> {
@@ -732,8 +734,8 @@ export class SocketManager {
             return;
         }
 
-        const recipient = room.getUserByUuid(recipientUuid);
-        if (recipient === undefined) {
+        const recipients = room.getUsersByUuid(recipientUuid);
+        if (recipients.length === 0) {
             console.error(
                 "In banUser, could not find user with id '" +
                     recipientUuid +
@@ -742,22 +744,24 @@ export class SocketManager {
             return;
         }
 
-        // Let's leave the room now.
-        room.leave(recipient);
+        for (const recipient of recipients) {
+            // Let's leave the room now.
+            room.leave(recipient);
 
-        const banUserMessage = new BanUserMessage();
-        banUserMessage.setMessage(message);
-        banUserMessage.setType("banned");
+            const banUserMessage = new BanUserMessage();
+            banUserMessage.setMessage(message);
+            banUserMessage.setType("banned");
 
-        const serverToClientMessage = new ServerToClientMessage();
-        serverToClientMessage.setBanusermessage(banUserMessage);
+            const serverToClientMessage = new ServerToClientMessage();
+            serverToClientMessage.setBanusermessage(banUserMessage);
 
-        // Let's close the connection when the user is banned.
-        recipient.socket.write(serverToClientMessage);
-        recipient.socket.end();
+            // Let's close the connection when the user is banned.
+            recipient.socket.write(serverToClientMessage);
+            recipient.socket.end();
+        }
     }
 
-    async sendAdminRoomMessage(roomId: string, message: string) {
+    async sendAdminRoomMessage(roomId: string, message: string, type: string) {
         const room = await this.roomsPromises.get(roomId);
         if (!room) {
             //todo: this should cause the http call to return a 500
@@ -772,7 +776,7 @@ export class SocketManager {
         room.getUsers().forEach((recipient) => {
             const sendUserMessage = new SendUserMessage();
             sendUserMessage.setMessage(message);
-            sendUserMessage.setType("message");
+            sendUserMessage.setType(type);
 
             const clientMessage = new ServerToClientMessage();
             clientMessage.setSendusermessage(sendUserMessage);
@@ -786,7 +790,7 @@ export class SocketManager {
         if (!room) {
             //todo: this should cause the http call to return a 500
             console.error(
-                "In sendAdminRoomMessage, could not find room with id '" +
+                "In dispatchWorldFullWarning, could not find room with id '" +
                     roomId +
                     "'. Maybe the room was closed a few milliseconds ago and there was a race condition?"
             );