+++ pusher/src/Services/SocketManager.ts
@@ -1,44 +1,45 @@
 import { PusherRoom } from "../Model/PusherRoom";
 import { CharacterLayer, ExSocketInterface } from "../Model/Websocket/ExSocketInterface";
 import {
+    AdminMessage,
+    AdminPusherToBackMessage,
+    AdminRoomMessage,
+    BanMessage,
+    CharacterLayerMessage,
+    EmoteEventMessage,
+    EmotePromptMessage,
     GroupDeleteMessage,
     ItemEventMessage,
+    JoinRoomMessage,
     PlayGlobalMessage,
+    PusherToBackMessage,
+    QueryJitsiJwtMessage,
+    RefreshRoomMessage,
+    ReportPlayerMessage,
     RoomJoinedMessage,
+    SendJitsiJwtMessage,
+    ServerToAdminClientMessage,
     ServerToClientMessage,
     SetPlayerDetailsMessage,
     SilentMessage,
     SubMessage,
-    ReportPlayerMessage,
+    UserJoinedRoomMessage,
     UserLeftMessage,
+    UserLeftRoomMessage,
     UserMovesMessage,
     ViewportMessage,
     WebRtcSignalToServerMessage,
-    QueryJitsiJwtMessage,
-    SendJitsiJwtMessage,
-    JoinRoomMessage,
-    CharacterLayerMessage,
-    PusherToBackMessage,
-    WorldFullMessage,
     WorldConnexionMessage,
-    AdminPusherToBackMessage,
-    ServerToAdminClientMessage,
-    EmoteEventMessage,
-    UserJoinedRoomMessage,
-    UserLeftRoomMessage,
-    AdminMessage,
-    BanMessage,
-    RefreshRoomMessage,
-    EmotePromptMessage,
+    TokenExpiredMessage,
     VariableMessage,
     ErrorMessage,
+    WorldFullMessage,
 } from "../Messages/generated/messages_pb";
 import { ProtobufUtils } from "../Model/Websocket/ProtobufUtils";
-import { ADMIN_API_URL, JITSI_ISS, SECRET_JITSI_KEY } from "../Enum/EnvironmentVariable";
+import { ADMIN_API_URL, JITSI_ISS, JITSI_URL, SECRET_JITSI_KEY } from "../Enum/EnvironmentVariable";
 import { adminApi } from "./AdminApi";
 import { emitInBatch } from "./IoSocketHelpers";
 import Jwt from "jsonwebtoken";
-import { JITSI_URL } from "../Enum/EnvironmentVariable";
 import { clientEventsEmitter } from "./ClientEventsEmitter";
 import { gaugeManager } from "./GaugeManager";
 import { apiClientRepository } from "./ApiClientRepository";
@@ -117,7 +118,7 @@ export class SocketManager implements ZoneEventListener {
                 console.warn("Admin connection lost to back server");
                 // Let's close the front connection if the back connection is closed. This way, we can retry connecting from the start.
                 if (!client.disconnecting) {
-                    this.closeWebsocketConnection(client, 1011, "Connection lost to back server");
+                    this.closeWebsocketConnection(client, 1011, "Admin Connection lost to back server");
                 }
                 console.log("A user left");
             })
@@ -138,24 +139,6 @@ export class SocketManager implements ZoneEventListener {
         if (socket.adminConnection) {
             socket.adminConnection.end();
         }
-    }
-
-    getAdminSocketDataFor(roomId: string): AdminSocketData {
-        throw new Error("Not reimplemented yet");
-        /*const data:AdminSocketData = {
-            rooms: {},
-            users: {},
-        }
-        const room = this.Worlds.get(roomId);
-        if (room === undefined) {
-            return data;
-        }
-        const users = room.getUsers();
-        data.rooms[roomId] = users.size;
-        users.forEach(user => {
-            data.users[user.uuid] = true
-        })
-        return data;*/
     }
 
     async handleJoinRoom(client: ExSocketInterface): Promise<void> {
@@ -406,17 +389,6 @@ export class SocketManager implements ZoneEventListener {
         }
     }
 
-    emitPlayGlobalMessage(client: ExSocketInterface, playglobalmessage: PlayGlobalMessage) {
-        if (!client.tags.includes("admin")) {
-            //In case of xss injection, we just kill the connection.
-            throw "Client is not an admin!";
-        }
-        const pusherToBackMessage = new PusherToBackMessage();
-        pusherToBackMessage.setPlayglobalmessage(playglobalmessage);
-
-        client.backConnection.write(pusherToBackMessage);
-    }
-
     public getWorlds(): Map<string, PusherRoom> {
         return this.rooms;
     }
@@ -463,7 +435,7 @@ export class SocketManager implements ZoneEventListener {
 
             client.send(serverToClientMessage.serializeBinary().buffer, true);
         } catch (e) {
-            console.error("An error occured while generating the Jitsi JWT token: ", e);
+            console.error("An error occurred while generating the Jitsi JWT token: ", e);
         }
     }
 
@@ -598,7 +570,20 @@ export class SocketManager implements ZoneEventListener {
         const serverToClientMessage = new ServerToClientMessage();
         serverToClientMessage.setWorldfullmessage(errorMessage);
 
-        client.send(serverToClientMessage.serializeBinary().buffer, true);
+        if (!client.disconnecting) {
+            client.send(serverToClientMessage.serializeBinary().buffer, true);
+        }
+    }
+
+    public emitTokenExpiredMessage(client: WebSocket) {
+        const errorMessage = new TokenExpiredMessage();
+
+        const serverToClientMessage = new ServerToClientMessage();
+        serverToClientMessage.setTokenexpiredmessage(errorMessage);
+
+        if (!client.disconnecting) {
+            client.send(serverToClientMessage.serializeBinary().buffer, true);
+        }
     }
 
     public emitConnexionErrorMessage(client: WebSocket, message: string) {
@@ -624,6 +609,36 @@ export class SocketManager implements ZoneEventListener {
         pusherToBackMessage.setEmotepromptmessage(emoteEventmessage);
 
         client.backConnection.write(pusherToBackMessage);
+    }
+
+    public async emitPlayGlobalMessage(
+        client: ExSocketInterface,
+        playGlobalMessageEvent: PlayGlobalMessage
+    ): Promise<void> {
+        if (!client.tags.includes("admin")) {
+            throw "Client is not an admin!";
+        }
+
+        const clientRoomUrl = client.roomId;
+        let tabUrlRooms: string[];
+
+        if (playGlobalMessageEvent.getBroadcasttoworld()) {
+            tabUrlRooms = await adminApi.getUrlRoomsFromSameWorld(clientRoomUrl);
+        } else {
+            tabUrlRooms = [clientRoomUrl];
+        }
+
+        const roomMessage = new AdminRoomMessage();
+        roomMessage.setMessage(playGlobalMessageEvent.getContent());
+        roomMessage.setType(playGlobalMessageEvent.getType());
+
+        for (const roomUrl of tabUrlRooms) {
+            const apiRoom = await apiClientRepository.getClient(roomUrl);
+            roomMessage.setRoomid(roomUrl);
+            apiRoom.sendAdminMessageToRoom(roomMessage, (response) => {
+                return;
+            });
+        }
     }
 }